[Chore] PaddingLabel설정값 DOOLabel로 합침(#48)
